임시 파일 수정3

+++ src/main/java/com/example/boardstudy2/Utils/FileUtil.java
@@ -1,53 +1,63 @@
 package com.example.boardstudy2.Utils;
 
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.PropertySource;
 import org.springframework.stereotype.Component;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.File;
+import java.io.FileOutputStream;
 import java.io.IOException;
+import java.io.InputStream;
+import java.nio.ByteBuffer;
+import java.nio.channels.FileChannel;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.nio.file.StandardCopyOption;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.UUID;
 
 @Component
+@PropertySource("classpath:application.properties")
 public class FileUtil {
 
+    // application.properties 임시 파일 경로 가져옴.
+    @Value("${spring.servlet.multipart.location}")
+    private String tempDir;
+
+    /**
+     * 파일 임시 저장
+     * @param files
+     * @return
+     * @throws IOException
+     */
     public List<String> tempFile(List<MultipartFile> files) throws IOException {
         List<String> fileNmList = new ArrayList<>();
-        String tempDir = "D:/Project/BoardStudy2/temp/";
 
         File dir = new File(tempDir);
-
-        // 폴더가 존재하면 파일 삭제 후 폴더도 삭제
-        File[] filesInDir = dir.listFiles();
-        if (filesInDir != null) {
-            for (File f : filesInDir) { // 폴더 내 파일 삭제
-                if (f.exists()) { // 파일 존재 여부 확인
-                    f.delete();
-                }
-            }
-        }
 
         // 폴더가 존재하지 않으면 새로 생성
         if(!dir.exists()){
             dir.mkdirs();
         }
 
-        // 파일 저장 및 파일 이름 추가
+        // 폴더가 존재하면 파일 삭제 후 폴더도 삭제
+        File[] filesInDir = dir.listFiles();
+        if (filesInDir != null) {
+            for (File f : filesInDir) { // 폴더 내 파일 삭제
+                if (f.exists() && f.isFile()) { // 파일 존재 여부 확인
+                    f.delete();  // 파일 삭제
+                }
+            }
+        }
+
+        // 파일 저장 및 파일 상태 검증
         for (MultipartFile file : files) {
             String originalFileName = file.getOriginalFilename();
-            if (originalFileName == null || originalFileName.isBlank()) {
-                throw new IOException("파일 이름이 올바르지 않습니다.");
-            }
 
             Path tempPath = Paths.get(tempDir, originalFileName);
-            byte[] bytes = file.getBytes();
-            Files.write(tempPath, bytes);
-
+            Files.copy(file.getInputStream(), tempPath, StandardCopyOption.REPLACE_EXISTING);
             fileNmList.add(originalFileName);
         }
